requirements.txt
argparse
colorlog
ConfigParser
cssselect
lxml
python-dateutil
requests
six
